Fire other spec unloading steps *after* pagehide
This change reorders two steps in the “unload a document” algorithm such
that the spec now requires UAs to wait to run any “unloading document
visibility change steps” from other applicable specifications (in particular,
the Page Visibility spec) until *after* firing the pagehide event.

In practice, in combination with requirements in the Page Visibility spec,
this change has the particular effect of stating that UAs must wait to fire
the visibilitychange event until after the pagehide event has fired. That
ordering matches the behavior implemented in Gecko and Blink, as well as the
WIP implementation in WebKit.

Otherwise, without this change, the combination of requirements with the
Page Visibility spec causes the specs together to mandate that the
visibilitychange event is fired *before* the pagehide event has fired
(which doesn’t match what implementations actually do).